+++ packages/administration/src/client-management/client-management-delete-dialog.component.html
