+++ packages/administration/src/client-management/client-management-delete-dialog.component.ts
@@ -2,7 +2,7 @@ import { Component, Input, OnInit } from '@angular/core';
 import { MatDialogRef } from '@angular/material/dialog';
 import { XmEventManager } from '@xm-ngx/core';
 
-import { Client, ClientService } from '../../shared';
+import { Client, ClientService } from '../../../../src/app/shared';
 
 @Component({
     selector: 'xm-client-mgmt-delete-dialog',
