+++ packages/administration/src/client-management/client-management-dialog.component.html
