+++ packages/administration/src/roles-management/roles-management-condition-dialog.component.html
