+++ packages/administration/src/roles-management/roles-management-detail.component.html
