+++ packages/administration/src/user-management/user-management-detail.component.html
