+++ packages/administration/src/user-management/user-management.component.ts
@@ -8,9 +8,9 @@ import { XmToasterService } from '@xm-ngx/toaster';
 import { JhiParseLinks } from 'ng-jhipster';
 import { finalize } from 'rxjs/operators';
 
-import { XM_EVENT_LIST } from '../../../app/xm.constants';
-import { Principal, RoleService, User, UserLogin, UserLoginService, UserService } from '../../shared';
-import { BaseAdminListComponent } from '../admin.service';
+import { XM_EVENT_LIST } from '../../../../src/app/xm.constants';
+import { Principal, RoleService, User, UserLogin, UserLoginService, UserService } from '../../../../src/app/shared';
+import { BaseAdminListComponent } from '../../../../src/app/admin/admin.service';
 import { UserLoginMgmtDialogComponent } from './user-login-management-dialog.component';
 import { UserMgmtDeleteDialogComponent } from './user-management-delete-dialog.component';
 import { UserMgmtDialogComponent } from './user-management-dialog.component';
