+++ src/app/admin/roles-management.route.ts
@@ -2,8 +2,8 @@ import { Injectable } from '@angular/core';
 import { CanActivate, Routes } from '@angular/router';
 
 import { Principal } from '@xm-ngx/core/auth';
-import { RoleMgmtDetailComponent } from './roles-management/roles-management-detail.component';
-import { RolesMgmtComponent } from './roles-management/roles-management.component';
+import { RoleMgmtDetailComponent } from '@xm-ngx/administration/roles-management/roles-management-detail.component';
+import { RolesMgmtComponent } from '@xm-ngx/administration/roles-management/roles-management.component';
 
 @Injectable()
 export class RolesResolve implements CanActivate {
